Separate out the CA scripts from client scripts
In a good world, you need the CA on a separate machine to the clients.
Therefore you should use the CA subdirectory on your CA machine and the
client subdirectory on any developer machines.

The sign_request.sh uses the CA key to sign incoming requests and
generate the signed certificates.